+++ backends/conrod_glium/Cargo.toml
@@ -26,3 +26,4 @@ find_folder = "0.3.0"
 image = "0.21"
 petgraph = "0.4"
 rand = "0.6"
+winit = "0.18"
