+++ backends/conrod_glium/examples/range_slider.rs
@@ -2,7 +2,7 @@
 
 #[macro_use] extern crate conrod_core;
 extern crate conrod_glium;
-extern crate conrod_winit;
+#[macro_use] extern crate conrod_winit;
 extern crate find_folder;
 extern crate glium;
 
@@ -57,7 +57,7 @@ fn main() {
         for event in event_loop.next(&mut events_loop) {
 
             // Use the `winit` backend feature to convert the winit event to a conrod one.
-            if let Some(event) = conrod_winit::convert_event(event.clone(), &display) {
+            if let Some(event) = support::convert_event(event.clone(), &display) {
                 ui.handle_event(event);
                 event_loop.needs_update();
             }
