+++ backends/conrod_winit/src/macros.rs
@@ -1,0 +1,345 @@
+//! A collection of macros for generating code related to winit+conrod interop.
+//!
+//! The reason we provide macros and don't implement functions using the `winit` crate directly is
+//! that conrod has many backends that use `winit`, often with differing versions. By providing
+//! macros, we allow these backends to generate code necessary for whatever version of winit they
+//! are currently using. This means we don't have to wait for all of the backend winit dependencies
+//! to synchronise before we can publish new conrod releases.
+
+/// Maps winit's key to a conrod `Key`.
+///
+/// Expects a `winit::VirtualKeyCode` as input and returns a `conrod_core::input::keyboard::Key`.
+///
+/// Requires that both the `winit` and `conrod_core` crates exist within the crate root.
+#[macro_export]
+macro_rules! convert_key {
+    ($keycode:expr) => {{
+        use conrod_core::input::keyboard::Key;
+
+        match $keycode {
+            winit::VirtualKeyCode::Key0 => Key::D0,
+            winit::VirtualKeyCode::Key1 => Key::D1,
+            winit::VirtualKeyCode::Key2 => Key::D2,
+            winit::VirtualKeyCode::Key3 => Key::D3,
+            winit::VirtualKeyCode::Key4 => Key::D4,
+            winit::VirtualKeyCode::Key5 => Key::D5,
+            winit::VirtualKeyCode::Key6 => Key::D6,
+            winit::VirtualKeyCode::Key7 => Key::D7,
+            winit::VirtualKeyCode::Key8 => Key::D8,
+            winit::VirtualKeyCode::Key9 => Key::D9,
+            winit::VirtualKeyCode::A => Key::A,
+            winit::VirtualKeyCode::B => Key::B,
+            winit::VirtualKeyCode::C => Key::C,
+            winit::VirtualKeyCode::D => Key::D,
+            winit::VirtualKeyCode::E => Key::E,
+            winit::VirtualKeyCode::F => Key::F,
+            winit::VirtualKeyCode::G => Key::G,
+            winit::VirtualKeyCode::H => Key::H,
+            winit::VirtualKeyCode::I => Key::I,
+            winit::VirtualKeyCode::J => Key::J,
+            winit::VirtualKeyCode::K => Key::K,
+            winit::VirtualKeyCode::L => Key::L,
+            winit::VirtualKeyCode::M => Key::M,
+            winit::VirtualKeyCode::N => Key::N,
+            winit::VirtualKeyCode::O => Key::O,
+            winit::VirtualKeyCode::P => Key::P,
+            winit::VirtualKeyCode::Q => Key::Q,
+            winit::VirtualKeyCode::R => Key::R,
+            winit::VirtualKeyCode::S => Key::S,
+            winit::VirtualKeyCode::T => Key::T,
+            winit::VirtualKeyCode::U => Key::U,
+            winit::VirtualKeyCode::V => Key::V,
+            winit::VirtualKeyCode::W => Key::W,
+            winit::VirtualKeyCode::X => Key::X,
+            winit::VirtualKeyCode::Y => Key::Y,
+            winit::VirtualKeyCode::Z => Key::Z,
+            winit::VirtualKeyCode::Apostrophe => Key::Unknown,
+            winit::VirtualKeyCode::Backslash => Key::Backslash,
+            winit::VirtualKeyCode::Back => Key::Backspace,
+            // K::CapsLock => Key::CapsLock,
+            winit::VirtualKeyCode::Delete => Key::Delete,
+            winit::VirtualKeyCode::Comma => Key::Comma,
+            winit::VirtualKeyCode::Down => Key::Down,
+            winit::VirtualKeyCode::End => Key::End,
+            winit::VirtualKeyCode::Return => Key::Return,
+            winit::VirtualKeyCode::Equals => Key::Equals,
+            winit::VirtualKeyCode::Escape => Key::Escape,
+            winit::VirtualKeyCode::F1 => Key::F1,
+            winit::VirtualKeyCode::F2 => Key::F2,
+            winit::VirtualKeyCode::F3 => Key::F3,
+            winit::VirtualKeyCode::F4 => Key::F4,
+            winit::VirtualKeyCode::F5 => Key::F5,
+            winit::VirtualKeyCode::F6 => Key::F6,
+            winit::VirtualKeyCode::F7 => Key::F7,
+            winit::VirtualKeyCode::F8 => Key::F8,
+            winit::VirtualKeyCode::F9 => Key::F9,
+            winit::VirtualKeyCode::F10 => Key::F10,
+            winit::VirtualKeyCode::F11 => Key::F11,
+            winit::VirtualKeyCode::F12 => Key::F12,
+            winit::VirtualKeyCode::F13 => Key::F13,
+            winit::VirtualKeyCode::F14 => Key::F14,
+            winit::VirtualKeyCode::F15 => Key::F15,
+            winit::VirtualKeyCode::Numpad0 => Key::NumPad0,
+            winit::VirtualKeyCode::Numpad1 => Key::NumPad1,
+            winit::VirtualKeyCode::Numpad2 => Key::NumPad2,
+            winit::VirtualKeyCode::Numpad3 => Key::NumPad3,
+            winit::VirtualKeyCode::Numpad4 => Key::NumPad4,
+            winit::VirtualKeyCode::Numpad5 => Key::NumPad5,
+            winit::VirtualKeyCode::Numpad6 => Key::NumPad6,
+            winit::VirtualKeyCode::Numpad7 => Key::NumPad7,
+            winit::VirtualKeyCode::Numpad8 => Key::NumPad8,
+            winit::VirtualKeyCode::Numpad9 => Key::NumPad9,
+            winit::VirtualKeyCode::NumpadComma => Key::NumPadDecimal,
+            winit::VirtualKeyCode::Divide => Key::NumPadDivide,
+            winit::VirtualKeyCode::Multiply => Key::NumPadMultiply,
+            winit::VirtualKeyCode::Subtract => Key::NumPadMinus,
+            winit::VirtualKeyCode::Add => Key::NumPadPlus,
+            winit::VirtualKeyCode::NumpadEnter => Key::NumPadEnter,
+            winit::VirtualKeyCode::NumpadEquals => Key::NumPadEquals,
+            winit::VirtualKeyCode::LShift => Key::LShift,
+            winit::VirtualKeyCode::LControl => Key::LCtrl,
+            winit::VirtualKeyCode::LAlt => Key::LAlt,
+            winit::VirtualKeyCode::RShift => Key::RShift,
+            winit::VirtualKeyCode::RControl => Key::RCtrl,
+            winit::VirtualKeyCode::RAlt => Key::RAlt,
+            winit::VirtualKeyCode::Home => Key::Home,
+            winit::VirtualKeyCode::Insert => Key::Insert,
+            winit::VirtualKeyCode::Left => Key::Left,
+            winit::VirtualKeyCode::LBracket => Key::LeftBracket,
+            winit::VirtualKeyCode::Minus => Key::Minus,
+            winit::VirtualKeyCode::Numlock => Key::NumLockClear,
+            winit::VirtualKeyCode::PageDown => Key::PageDown,
+            winit::VirtualKeyCode::PageUp => Key::PageUp,
+            winit::VirtualKeyCode::Pause => Key::Pause,
+            winit::VirtualKeyCode::Period => Key::Period,
+            winit::VirtualKeyCode::Right => Key::Right,
+            winit::VirtualKeyCode::RBracket => Key::RightBracket,
+            winit::VirtualKeyCode::Semicolon => Key::Semicolon,
+            winit::VirtualKeyCode::Slash => Key::Slash,
+            winit::VirtualKeyCode::Space => Key::Space,
+            winit::VirtualKeyCode::Tab => Key::Tab,
+            winit::VirtualKeyCode::Up => Key::Up,
+            _ => Key::Unknown,
+        }
+    }};
+}
+
+/// Maps winit's mouse button to conrod's mouse button.
+///
+/// Expects a `winit::MouseButton` as input and returns a `conrod_core::input::MouseButton` as
+/// output.
+///
+/// Requires that both the `conrod_core` and `winit` crates are in the crate root.
+#[macro_export]
+macro_rules! convert_mouse_button {
+    ($mouse_button:expr) => {{
+        use conrod_core::input::MouseButton;
+        match $mouse_button {
+            winit::MouseButton::Left => MouseButton::Left,
+            winit::MouseButton::Right => MouseButton::Right,
+            winit::MouseButton::Middle => MouseButton::Middle,
+            winit::MouseButton::Other(0) => MouseButton::X1,
+            winit::MouseButton::Other(1) => MouseButton::X2,
+            winit::MouseButton::Other(2) => MouseButton::Button6,
+            winit::MouseButton::Other(3) => MouseButton::Button7,
+            winit::MouseButton::Other(4) => MouseButton::Button8,
+            _ => MouseButton::Unknown,
+        }
+    }};
+}
+
+/// A macro for converting a `winit::WindowEvent` to a `Option<conrod_core::event::Input>`.
+///
+/// Expects a `winit::WindowEvent` and a reference to a window implementing `WinitWindow`.
+/// Returns an `Option<conrod_core::event::Input>`.
+#[macro_export]
+macro_rules! convert_window_event {
+    ($event:expr, $window:expr) => {{
+        // The window size in points.
+        let (win_w, win_h) = match $window.get_inner_size() {
+            Some((w, h)) => (w as conrod_core::Scalar, h as conrod_core::Scalar),
+            None => return None,
+        };
+
+        // Translate the coordinates from top-left-origin-with-y-down to centre-origin-with-y-up.
+        let tx = |x: conrod_core::Scalar| x - win_w / 2.0;
+        let ty = |y: conrod_core::Scalar| -(y - win_h / 2.0);
+
+        // Functions for converting keys and mouse buttons.
+        let map_key = |key| convert_key!(key);
+        let map_mouse = |button| convert_mouse_button!(button);
+
+        match $event {
+            winit::WindowEvent::Resized(winit::dpi::LogicalSize { width, height }) => {
+                Some(conrod_core::event::Input::Resize(width as _, height as _).into())
+            },
+
+            winit::WindowEvent::ReceivedCharacter(ch) => {
+                let string = match ch {
+                    // Ignore control characters and return ascii for Text event (like sdl2).
+                    '\u{7f}' | // Delete
+                    '\u{1b}' | // Escape
+                    '\u{8}'  | // Backspace
+                    '\r' | '\n' | '\t' => "".to_string(),
+                    _ => ch.to_string()
+                };
+                Some(conrod_core::event::Input::Text(string).into())
+            },
+
+            winit::WindowEvent::Focused(focused) =>
+                Some(conrod_core::event::Input::Focus(focused).into()),
+
+            winit::WindowEvent::KeyboardInput { input, .. } => {
+                input.virtual_keycode.map(|key| {
+                    match input.state {
+                        winit::ElementState::Pressed =>
+                            conrod_core::event::Input::Press(conrod_core::input::Button::Keyboard(map_key(key))).into(),
+                        winit::ElementState::Released =>
+                            conrod_core::event::Input::Release(conrod_core::input::Button::Keyboard(map_key(key))).into(),
+                    }
+                })
+            },
+
+            winit::WindowEvent::Touch(winit::Touch { phase, location, id, .. }) => {
+                let winit::dpi::LogicalPosition { x, y } = location;
+                let phase = match phase {
+                    winit::TouchPhase::Started => conrod_core::input::touch::Phase::Start,
+                    winit::TouchPhase::Moved => conrod_core::input::touch::Phase::Move,
+                    winit::TouchPhase::Cancelled => conrod_core::input::touch::Phase::Cancel,
+                    winit::TouchPhase::Ended => conrod_core::input::touch::Phase::End,
+                };
+                let xy = [tx(x), ty(y)];
+                let id = conrod_core::input::touch::Id::new(id);
+                let touch = conrod_core::input::Touch { phase: phase, id: id, xy: xy };
+                Some(conrod_core::event::Input::Touch(touch).into())
+            }
+
+            winit::WindowEvent::CursorMoved { position, .. } => {
+                let winit::dpi::LogicalPosition { x, y } = position;
+                let x = tx(x as conrod_core::Scalar);
+                let y = ty(y as conrod_core::Scalar);
+                let motion = conrod_core::input::Motion::MouseCursor { x: x, y: y };
+                Some(conrod_core::event::Input::Motion(motion).into())
+            },
+
+            winit::WindowEvent::MouseWheel { delta, .. } => match delta {
+                winit::MouseScrollDelta::PixelDelta(winit::dpi::LogicalPosition { x, y }) => {
+                    let x = x as conrod_core::Scalar;
+                    let y = -y as conrod_core::Scalar;
+                    let motion = conrod_core::input::Motion::Scroll { x: x, y: y };
+                    Some(conrod_core::event::Input::Motion(motion).into())
+                },
+
+                winit::MouseScrollDelta::LineDelta(x, y) => {
+                    // This should be configurable (we should provide a LineDelta event to allow for this).
+                    const ARBITRARY_POINTS_PER_LINE_FACTOR: conrod_core::Scalar = 10.0;
+                    let x = ARBITRARY_POINTS_PER_LINE_FACTOR * x as conrod_core::Scalar;
+                    let y = ARBITRARY_POINTS_PER_LINE_FACTOR * -y as conrod_core::Scalar;
+                    Some(conrod_core::event::Input::Motion(conrod_core::input::Motion::Scroll { x: x, y: y }).into())
+                },
+            },
+
+            winit::WindowEvent::MouseInput { state, button, .. } => match state {
+                winit::ElementState::Pressed =>
+                    Some(conrod_core::event::Input::Press(conrod_core::input::Button::Mouse(map_mouse(button))).into()),
+                winit::ElementState::Released =>
+                    Some(conrod_core::event::Input::Release(conrod_core::input::Button::Mouse(map_mouse(button))).into()),
+            },
+
+            winit::WindowEvent::Refresh => {
+                Some(conrod_core::event::Input::Redraw)
+            },
+
+            _ => None,
+        }
+    }};
+}
+
+/// A macro for converting a `winit::Event` to a `conrod_core::event::Input`.
+///
+/// Expects a `winit::Event` and a reference to a window implementing `WinitWindow`.
+/// Returns an `Option<conrod_core::event::Input>`.
+///
+/// Invocations of this macro require that a version of the `winit` and `conrod_core` crates are
+/// available in the crate root.
+#[macro_export]
+macro_rules! convert_event {
+    ($event:expr, $window:expr) => {{
+        match $event {
+            winit::Event::WindowEvent { event, .. } => convert_window_event!(event, $window),
+            _ => None,
+        }
+    }};
+}
+
+/// Convert a given conrod mouse cursor to the corresponding winit cursor type.
+///
+/// Expects a `conrod_core::cursor::MouseCursor`, returns a `winit::MouseCursor`.
+///
+/// Requires that both the `conrod_core` and `winit` crates are in the crate root.
+#[macro_export]
+macro_rules! convert_mouse_cursor {
+    ($cursor:expr) => {{
+        use conrod_core::cursor;
+        match $cursor {
+            cursor::MouseCursor::Text => winit::MouseCursor::Text,
+            cursor::MouseCursor::VerticalText => winit::MouseCursor::VerticalText,
+            cursor::MouseCursor::Hand => winit::MouseCursor::Hand,
+            cursor::MouseCursor::Grab => winit::MouseCursor::Grab,
+            cursor::MouseCursor::Grabbing => winit::MouseCursor::Grabbing,
+            cursor::MouseCursor::ResizeVertical => winit::MouseCursor::NsResize,
+            cursor::MouseCursor::ResizeHorizontal => winit::MouseCursor::EwResize,
+            cursor::MouseCursor::ResizeTopLeftBottomRight => winit::MouseCursor::NwseResize,
+            cursor::MouseCursor::ResizeTopRightBottomLeft => winit::MouseCursor::NeswResize,
+            _ => winit::MouseCursor::Arrow,
+        }
+    }};
+}
+
+/// Generate a set of conversion functions for converting between types of the crate's versions of
+/// `winit` and `conrod_core`.
+#[macro_export]
+macro_rules! conversion_fns {
+    () => {
+        /// Convert a `winit::VirtualKeyCode` to a `conrod_core::input::keyboard::Key`.
+        pub fn convert_key(keycode: winit::VirtualKeyCode) -> conrod_core::input::keyboard::Key {
+            convert_key!(keycode)
+        }
+
+        /// Convert a `winit::MouseButton` to a `conrod_core::input::MouseButton`.
+        pub fn convert_mouse_button(
+            mouse_button: winit::MouseButton,
+        ) -> conrod_core::input::MouseButton {
+            convert_mouse_button!(mouse_button)
+        }
+
+        /// Convert a given conrod mouse cursor to the corresponding winit cursor type.
+        pub fn convert_mouse_cursor(
+            cursor: conrod_core::cursor::MouseCursor,
+        ) -> winit::MouseCursor {
+            convert_mouse_cursor!(cursor)
+        }
+
+        /// A function for converting a `winit::WindowEvent` to a `conrod_core::event::Input`.
+        pub fn convert_window_event<W>(
+            event: winit::WindowEvent,
+            window: &W,
+        ) -> Option<conrod_core::event::Input>
+        where
+            W: $crate::WinitWindow,
+        {
+            convert_window_event!(event, window)
+        }
+
+        /// A function for converting a `winit::Event` to a `conrod_core::event::Input`.
+        pub fn convert_event<W>(
+            event: winit::Event,
+            window: &W,
+        ) -> Option<conrod_core::event::Input>
+        where
+            W: $crate::WinitWindow,
+        {
+            convert_event!(event, window)
+        }
+    };
+}
